xiaohongshu-hub/pyproject.toml
[project]
name = "xiaohongshu-hub"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pycryptodome>=3.20",
]
